requirements.txt
Pillow
numpy
scikit-image
opencv-python